Prefix sum example
This is a straightforward implementation of decoupled lookback. It's
very rough in a number of ways - lots of hardwired constants (it won't
work unless workgroup size is square of subgroup), non-parallel
implementation of lookback. Also, not really suited to performance
evaluation because it works in host-coherent buffers; device-local would
probably be lots faster.