Cut down winding lookups more aggressively
Inspired by the original kernel4 code, this change avoids the expensive lookup whenever it's determined that the lookup result would be discarded (bit AND against a zero).

It's unfortunate that we end up doing a lot of redundant work in kernel4, but at least we gain up to 10% of improvement for now.