Add coherent case
Add a test with a `coherent` qualifier on the buffer and explicit memory
barriers. This is probably the most interesting case as it should be
portable.

Interestingly, tests on AMD 5700XT show that the explicit memory model
has a nontrivial performance edge, so it might be worth reaching for
that when it's available.