Fix shader hang in coarse path
I'm observing a hang in the coarse path rasterizer on Nvidia 1060 (not
on other GPUs, and I think only after a more recent driver update).
Investigation shows that the hang happens when the loop to output line
segments from flattening does not terminate. I believe this is a
floating point inconsistency of some kind - the "target" should always
be at least the sum of the "val" for all quadratics, but if it
undershoots the loop may continue when it should not.

It's *possible* this patch is masking a deeper logic error. Ideally, I'd
instrument the shader and capture a trace of when exactly this goes
wrong. But in any case this patch shouldn't make things any worse.