Alternative strategies for elements barrier
This patch provides an #ifdef METAL for switching between two different
strategies for the barriers in the elements shader. With it enabled, the
barriers are in workgroup uniform control flow, which is compatible with
translation to Metal, but fails to compile in FXC, as FXC's uniformity
analysis fails. With it disabled, the lookback logic (including
barriers) runs only in a single thread, which compiles in FXC (though
not tested), but creates problems in Metal.

In testing on Android, the METAL version seems slightly faster. On AMD
5700 XT, there is no measurable difference.

I'm inclined not to commit this, but it's potentially useful if we want
to explore cs_5_0 compatibility.