Adaptation of MotionMark 1.2 path test
This is a quick and dirty adapation of the MotionMark 1.2 path test into
Rust. It is not at all a fair comparison, as the original is using miter
joins and butt caps, while this one is round/round (to use the SDF-based
stroke renderer). Nonetheless, it might be fun to play with.