Hierarchical fine raster
This commit introduces a hierarchical approach to exploit sparsity in fine rasterization.

A cheap and non-antialiased is performed first for each pixel, then a sparse AA pass is performed for only the tiles that are touched by the line segment.

The performance is slightly worse (~10% slower k4) than the previous approach, unfortunately. While this approach doesn't give much benefit as-is, it can be effective in bringing down the cost of a bitmask-based supersampling approach, which notably solves conflation artifacts.