Don't encode anything for "empty" brushes
I was crashing on ramp generation when a gradient contained an empty color stop collection. This changes `SceneBuilder` to check for that case and prevent encoding anything at all. Does the same for solid brushes with alpha of 0 because might as well.

Also adds a new method for creating a surface with a specific presentation mode to `util::RenderContext`. This is a new method rather than adding a parameter to prevent breaking xilem.