Delete load balanced option for path_count
It adds complexity and has no benefit. We need to do more aggressive path culling in path_count, so keeping things simple is a win.

This is something we might revisit later. Special-casing one-tile lines is probably a win (and this case can be detected in flattening). Then load balancing *may* be a win for the remaning lines, but that is not clear.