Clip to bounding box in path_count
This commit clips the line to the bounding box in the path_count stage, only iterating tiles within the bounding box.

As a consequence, the path_tiling stage need not deal with invalid SegmentCount values.

The clipping technique is non-obvious. Usually the coordinates are clipped against the edges of the bounding box. However, here the goal is to determine the contiguous range of indices that fall inside the bbox.

Care is taken to be numerically robust. A linear solve (division) is used to find an index that is close to the boundary (either i or i+1 is correct). Then the index-to-coordinate logic is used to decide between those two choices.

An additional wrinkle is bumping backdrop when all or a portion of the line is to the left of the bounding box. This is done in a separate loop.