Another checkpoint
This commit wires up CPU and GPU handling of resources, passing the former to the shader function. The main thing missing is deferred upload of buffers so the output of a CPU stage can be used as the input to GPU.

Again it does not regress a GPU-only pipeline. There is some refactoring of external resources, as the idea of a transient resource map unifies an external resource and a buffer upload.