Checkpoint
A few simple CPU shaders work; flatten doesn't. Checkpoint before starting work on that.

The immediate problem is that Clear materializes the buffer (in this case, bump) on the GPU. The best strategy would be to defer the materialization (and thus the clear), as it's hard to know whether it should be CPU or GPU.