Checkpoint with clipping
This commit adds clipping to the CPU shader pipeline, but there are artifacts in the cardioid test scene.

Detective work is needed, as these artifacts seem to be present in the GPU pipeline as well (though not necessarily the same). A clue: the artifacts are flickery with coarse on GPU, stable with coarse on CPU.