Fix zero-clip artifact
Tiles that don't make it through coarse rasterization because of the zero-clip optimization must be fully ignored. Previously the count of such tiles was being misinterpreted as an index for the allocation slice, causing valid data to be overwritten.

This patch uses the sign bit to disambiguate the two cases, and explicitly discards tiles with the wrong sign.