[shaders] Remove unused resource bindings
clip_reduce, path_coarse_full, and fine stages declare some resource
bindings that are not used.

These handful of bindings have now been removed and the binding indices
were rearranged to be compact.

This partially helps with the bug reported in #330. The actual fix for
that issue actually makes it so that the unused bindings are no longer
a problem with native integrations of the vello_shader crate, but this
is a reasonable clean up nontheless.