[shaders][msl] Enable `zero_initialize_workgroup_memory` setting
Zero-initializing workgroup memory works around a Metal driver bug
that can cause the tile_alloc and coarse stages to freeze up the system
when running natively on Metal due to their use of `workgroupUniformLoad`.
See gfx-rs/naga#2482 for more details about this bug.

Zero-initializing workgroup memory also makes the native Metal shaders
match the behavior of the current wgpu runner and the invariants of the
original WGSL that the shaders were authored in.