Add CPU shaders
This commit adds a full pipeline of CPU shaders for the compute stage of the pipeline (not counting fine rasterization). It's based on what's in the multi branch, but also ports bounding box fixes and other cleanups.

Currently the selection of shaders is all or nothing (based on the use-cpu flag), but will probably become more fine-grained.

Going forward, the intent is to keep CPU and GPU shaders in sync. It has been quite valuable in diagnosing problems in the pipeline.