[encoding] Do not allow zero-length buffer sizes
It is possible for a BufferSize of 0 bytes to trigger validation
warnings as such buffers still need to get bound to their pipeline. The
standard vello engine (src/engine.rs) handles this when a BufProxy is
requested by setting the requested size to at least 16.

This is now handled by the encoding crate by selecting a buffer size
that can fit at least one element.