[test_scenes} Stroke cap variants of longpathdash
Added a butt cap and round cap variant of the pathological longpathdash
test case. The round caps exhibit the expected increase in the number of
line segments post-flattening, and hence a significant increase in
required GPU memory.

Also added a new `scene!` macro variant for custom test names and
factory functions that return a test function.