Prototype 8 bit winding number accumulation
Changes the winding number accumulation in msaa8 mode to 8 bits per sample. This is prototype code and currently breaks the msaa16 mode; it is intended to diagnose whether the artifacts are strictly due to overflow, and to point the way to a real implementation.

Prefix sums in both x and y direction are a little cleaner, avoiding a race (not UB because it's atomics).