Shitty offset curves for GPU stroke scaffolding
Implemented a stroking scheme in the flatten stage in which every
flattened line segment gets offset along the curve normal by the line
width. The line vertices are offset _after_ subdivision and does not
try to accurately estimate subdivisions for the offset curve.

Cusps are not detected and are prone to failure. Regions with high
curvature (especially around self-intersections) are handled by simply
boosting the subdivision count using the curve offset which is not
correct but avoids awfully missing samples when the tolerance is set too
low.

This is intended as a placeholder but is the first step towards
implementing correct strokes without optimizing for performance.