[flatten] Implement round join style
Implemented the logic for round join and cap styles. The logic generates
lines directly from a circular arc. The arc gets flattened in the
curve's local coordinate space and transformed to device-space
coordinates post-flattening (just like strokes work in `flatten_arc`).

An advantage of flattening an arc directly is that the logic is very
simple and involves way less ALU compared to `flatten_cubic`. The
disadvantage is that the structure of join/cap style handling is now
highly divergent.

To reduce divergence, the shader could be restructured to model a (not
perfectly circular) arc with cubic Beziers and let the control flow
converge at a call to `flatten_cubic`.