[encoding] Do not encode zero-length path segments
Zero-length segments lead to numerical errors and lead to extra
complexity and wasted work (in the form of thread allocation) for GPU
flattening. We now detect and drop them at encoding time.

Currently, a path segment is considered zero-length if all of its
control points (in local coordinates) are within EPSILON (1e-12) of
each other. This may not be the desired behavior under transform.

Also since the length here isn't in terms of the actual arc length,
this won't detect all curves that are within an EPSILON sized bounding
box (which may have a larger distance between their control points).

For stroking purposes, the distance between the control points is what
matters most as that's what's used to compute a curve's tangent.