debug_layers feature
This introduces a new debug module and DebugLayers data structure that
can render debug visualizations of GPU buffers that are internal to the
vello pipeline. This currently supports a line-based visualization of
the LineSoup and PathBbox buffers.

The debug visualization depends on CPU read-back of the BumpAllocators
buffer to issue a draw call. This could technically be avoided with an
indirect draw but the visualizations will eventually include CPU-side
processing and validation.

The draws are recorded to the same Recording as the blit in
`render_to_texture_async`. The buffers that are used by the draw
commands are temporarily retained outside of the `Render` data
structure. These buffers are currently released back to the engine
explicitly and in various places in code since safe resource removal
currently requires a Recording.