[shaders] Don't apply force_rw_storage to fine
The force_rw_storage feature is a hack that enables a renderer to bind
suballocations from the same buffer object to both readonly and
read-write bindings inside a pipeline layout.

The AA variants of the fine stage don't compile with this feature
enabled, since the read-write PTCL buffer results in a uniformity
analysis failure.

Address this by excluding the `fine` stage from the force_rw_storage
feature to preserve the shader's correctness. Since all storage bindings
in `fine` are readonly, there is no restriction against binding
suballocations to any of the fine pipeline variants in the first place.