Checkpoint of Euler spiral based stroke expansion (#496)
* Checkpoint euler patch

This just imports Arman's patch and tries to update it for other changes.

* Checkpoint numerical robustness work

Apply Euler spiral numerical robustness to CPU shader.

There are a number of minor tweaks, and some important changes:

* Cubic segments are computed with points and derivatives, not subdivision
* Those derivatives are adjusted if they are near-zero
* The ESPC integral is robust to small k1 and dist

Note that dealing with small dist is essential for doing ES flattening for fills as well as strokes.

* Checkpoint

This pipes through the start/end points (always using them as endpoints) and always uses Euler flattening.

It seems to basically work for fills, but has lots of artifacts still for strokes.

* Don't special case lines

Drop zero-length lines. This doesn't get rid of all the code, but it shouldn't be needed for correctness.

In addition, zero length lines are dropped. We need to come back to this, they do have some semantics for strokes.

* Force style for each encoded glyph

* Better continuity of flattening

Reuse endpoints of Euler segments. In theory, this should ensure watertight meshes, but we're definitely not seeing it.

* Fix end normals

Calculation of end normals for stroke offset was mistaken. At this point, it's starting to look fairly good. Some of the tricky strokes need investigation, and handling of zero-length lines is still a bit dodgy.

* Checkpoint WGSL work

Euler flattening seems to be almost working as long as subdivision is disabled, at least for fills. This commit has that disabled.

* Fix bugs

Simple typo in subdivision logic (and re-enable subdivision, which was disabled in previous commit).

Sign wrong in inverse integral.

Typo in calculation of start tangent (repeated p3 twice instead of p2, p3).

Note: also enables GPU-side stroking

* Small fixes and cleanups

Make handling of small tangent consistent. Remove cubic flattening. Make subdivision limiting logic of CPU shader consistent with GPU version. Get rid of unused is_line predicate on CPU side.

* Clippy fixes

* Fix copyright headers

(and I also added this to my local pre-push git hook)

* Review suggestions

Mostly doc improvements, but also improvements to robustness for short chords.

* Tune tolerance value

I have validated that the tolerance value for both arcs and Bézier segments is calibrated correctly. Thus, it makes sense to set both to a value of 0.25, so that caps are consistent. This comes quite close to matching the amount of subdivision in current main, and the line count is also consistent between CPU and GPU stroke expansion.

* Make GPU tolerance consistent with CPU

Oops, I made the change only on the CPU side. This brings GPU into agreement.

* Address review comments

A bit more comment, but also a more robust protection against a divide by zero, which happens in the colinear double-cusp case, as exercised by tricky-strokes.