Rework robustness of cubic params
This commit contains a somewhat different approach to robust cubic
params than #520. When the chord length is short, propose a straight
line, and then calculate a good estimate of the actual error of the
cubic relative to that line.

More changes are made in downstream logic to avoid dividing by a zero
chord length. The effective chord length computed in cubic params is now
at least DIST_THRESH.

Also contains simplification of control flow in flatten_euler which
happens to work around miscompilation on some AMD cards.

WIP, logic is GPU-only.