Apply robustness fixes to CPU shader as well
This should make CPU and GPU shaders consistent. Some ideas were borrowed from #520, in particular the divide by zero of the chord length in offset calculation is avoided (though there is now refactoring so the robustness logic is applied once and the effective chord length is now stored in cubic_params).

The rustdoc for CubicParams::from_points_derivs also documents a bit of the rationale for the current robustness logic.