[encoding] Inflate estimates for offset curves
- Added a highly crude scale to segment and line counts in the presence
  of a non-zero offset (i.e. stroked curves). This is done to account
  for the inflated flattening of an offset curve around high-curvature
  zones. I believe this could get refined but that requires some work
  to find the best way to analyze the complexity of the curve with few
  CPU cycles.

- Fixed a bug in join calculation that was wrongfully incrementing the
  line count instead of segments.

This makes the estimate robust for tricky strokes but also increases
the margin of the estimate over the true bump counts. I expect some of
this will improve when the estimation moves to resolve time, such that
the crude bloating of the counts due to any scale applied to a scene
fragment will no longer be necessary.