Rework robustness of cubic params (#521)
* Rework robustness of cubic params

This commit contains a somewhat different approach to robust cubic
params than #520. When the chord length is short, propose a straight
line, and then calculate a good estimate of the actual error of the
cubic relative to that line.

More changes are made in downstream logic to avoid dividing by a zero
chord length. The effective chord length computed in cubic params is now
at least DIST_THRESH.

Also contains simplification of control flow in flatten_euler which
happens to work around miscompilation on some AMD cards.

WIP, logic is GPU-only.

* Apply robustness fixes to CPU shader as well

This should make CPU and GPU shaders consistent. Some ideas were borrowed from #520, in particular the divide by zero of the chord length in offset calculation is avoided (though there is now refactoring so the robustness logic is applied once and the effective chord length is now stored in cubic_params).

The rustdoc for CubicParams::from_points_derivs also documents a bit of the rationale for the current robustness logic.

* Address review feedback

Mostly added comments and a bit of renaming, no logic changes.